re_rar_set: Avoid buffer overread and probably fix big-endian bug
eaddr is a byte buffer of length ETHER_ADDR_LEN, i.e. 6 bytes, so we
shouldn't try to read bytes at offset 6 and 7. This also looks like it
will end up using the wrong half of the word (the two following bytes)
on big-endian systems. Fix this by just reading 2 bytes, as is done
below for some of the other CSRs.